Remove hostNetwork:true from rproxy
This change is quite subtle and needs some explaination.

Firstly, the deployment spec already lists a hostPort, when
hostNetwork:true, this does nothing.

When hostNetwork:false, docker-proxy listens on these ports
and proxies (via userspace) packets to the containerPort
within the Pod's network namespace.

Consequence 1: containers can bind to a different port than
what port is accepting external traffic.

Consequence 2: There may be a performance hit as packets
are proxied through a separate process, and is subject to
any bugs/quirks in that process/TCP stack.
This may also have an impact on debugability.

Setting hostNetwork:false runs the Pod inside it's own
network namespace.

Consequence 3: This adds to the container's security,
separating which connections it can make and which
external ports processes in the container can bind to.

Also note that a NodePort Service cannot be used to
achieve this same result. NodePort, unlike HostPort
cannot bind to low/privileged ports by default.