Add metallb for Service(type=LoadBalancer)
This adds [metallb](https://metallb.universe.tf) definitions.

MetalLB is packaged as a `controller` (Deployment) and a `speaker`
(DaemonSet) that listens for `Service`s that request
`type=LoadBalancer`.

The committed config file runs the "LoadBalancer" in ARP mode, which
advertises an IP in the configured pool to the local network segment.

This means that, any host in the LAN (we're using RFC1918 addresses)
can ARP for IPs managed by the kubernetes cluster. In ARP mode, only
one node currently listens for ARP requests (using leadership election)
and all connections for `Service`s will be pinned through this node.